pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools_scm[toml]>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "kledo-api-mcp"
version = "0.2.0-beta"
description = "Model Context Protocol server for Kledo accounting software"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Kevin Zakaria", email = "kledo-api-mcp@users.noreply.github.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business :: Financial :: Accounting",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
keywords = ["mcp", "kledo", "accounting", "erp", "api", "claude", "ai", "automation"]
requires-python = ">=3.11"
dependencies = [
    "mcp>=0.9.0",
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "loguru>=0.7.0",
    "python-dateutil>=2.8.0",
    "rapidfuzz>=3.14.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "erdantic>=1.0.0",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.24.0",
]
docs = [
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.24.0",
]

[project.scripts]
kledo-mcp = "kledo_mcp:main"

[project.urls]
Homepage = "https://github.com/efacsen/kledo-api-mcp"
Documentation = "https://efacsen.github.io/kledo-api-mcp/"
Repository = "https://github.com/efacsen/kledo-api-mcp.git"
Issues = "https://github.com/efacsen/kledo-api-mcp/issues"
Changelog = "https://github.com/efacsen/kledo-api-mcp/blob/main/MCP_SERVER_UPDATES.md"

[tool.setuptools]
py-modules = ["kledo_mcp"]
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
src = ["*.yaml", "*.yml", "*.json", "*.md"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # Line too long
    "B008",  # Do not perform function calls in argument defaults
    "C901",  # Too complex
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true

[[tool.mypy.overrides]]
module = [
    "mcp.*",
    "httpx.*",
    "loguru.*",
    "pydantic.*",
    "rapidfuzz.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "tests.py"]
python_classes = ["Test*", "*Tests"]
python_functions = ["test_*", "*_tests"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--disable-warnings",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html:htmlcov",
    "--cov-fail-under=45",
    "--cov-config=.coveragerc",
]
asyncio_mode = "auto"
